Update DIRA_intro.md
Unfortunately, the Federalist build does not support <li></li> or <ul><li></li></ul> coding for bullets in tables. That coding renders bullets perfectly in md tables in md Preview mode, but the HTML coding, not the bullets, appear in the Federalist build. I put in asterisks for now in Table 6. It's not the best aesthetic solution, but it delineates bulleted items for now in that table.